+++ examples/Cavity2D/libudf/lnamd64/2ddp_node/fsi.h
@@ -1,1 +1,12 @@
-../../src/fsi.h
+#ifndef FSI_H_
+#define FSI_H_
+
+#include "udf.h"
+#include "dynamesh_tools.h"
+
+void fsi_init(Domain* domain);
+void fsi_write_and_advance();
+void fsi_grid_motion(Domain* domain, Dynamic_Thread* dt, real time, real dtime);
+void fsi_plot_coords();
+
+#endif /* FSI_H_ */
