--- examples/Cavity2D/libudf/lnamd64/2ddp_node/fsi_extrafunctions.c
@@ -1,1 +1,0 @@
-../../src/fsi_extrafunctions.c
