--- examples/Cavity2D/libudf/lnamd64/2ddp_node/train_udf.c
@@ -1,1 +1,0 @@
-../../src/train_udf.c
