+++ examples/Cavity2D/libudf/lnamd64/2ddp_node/user.udf
@@ -1,5 +1,5 @@
 CSOURCES= fsi_udf.c fsi.c
 HSOURCES= fsi.h
-FLUENT_INC=/home/ishaan/ansys_inc/v195/fluent
+FLUENT_INC= /opt/Software/ansys/v202/fluent
 HOST_FLAGS=-DUSE_UDF_HOST_DUMMY
 GPU_SUPPORT=off
