--- examples/Cavity2D/libudf/lnamd64/2ddp_node/wave_maker_udf.c
@@ -1,1 +1,0 @@
-../../src/wave_maker_udf.c
