--- examples/Cavity2D/libudf/lnamd64/2ddp_node/wave_profile.c
@@ -1,1 +1,0 @@
-../../src/wave_profile.c
