--- examples/Cavity2D/libudf/src/wave_profile_udf.h
@@ -1,10 +1,0 @@
-#ifdef TEST
-#define real double
-#endif
-
-real get_ratio_triangle_is_covered (
-  real coords[6],
-  real wave_height );
-
-real compute_triangle_area_2d(
-  real coords[6]);
