+++ examples/Cavity2D/precice-config.xml
@@ -5,6 +5,7 @@
   <log enabled="true">
     <sink type="stream" output="stdout"  filter= "%Severity% > trace" format="preCICE:%ColorizedSeverity% %Message%" enabled="true"/>	
     <sink type="file" output="fluentCSM_debug.log" filter="%Severity% >= trace" enabled="true"/>
+    <sink type="file" output="fluentCSM_alldebug.log" filter="" enabled="true"/>
   </log>
 
   <solver-interface dimensions="2">
@@ -27,21 +28,25 @@
       <use-mesh name="beam" from="CSMdummy"/>
       <write-data name="Forces"     mesh="moving_base"/>
       <read-data  name="Displacements" mesh="moving_base"/>
-      <mapping:nearest-projection   direction="write" from="moving_base" to="beam" constraint="consistent"/>
+      <mapping:nearest-projection   direction="write" from="moving_base" to="beam" constraint="conservative"/>
+      <!--
       <mapping:nearest-projection   direction="read"  from="beam" to="moving_base" constraint="consistent"/>
+      -->
     </participant>
 
     <participant name="CSMdummy">
       <use-mesh name="beam" provide="yes"/>
+      <use-mesh name="moving_base" from="Fluent"/>
       <write-data name="Displacements" mesh="beam"/>
       <read-data  name="Forces"     mesh="beam"/>
+      <mapping:nearest-projection   direction="write" from="beam" to="moving_base" constraint="conservative"/>
     </participant>
 
     <m2n:sockets from="Fluent" to="CSMdummy"/>
 
     <coupling-scheme:serial-implicit> 
       <participants first="Fluent" second="CSMdummy"/> 
-      <time-window-size method="first-participant"/>
+      <time-window-size value="0.001"/>
       <max-iterations value="100"/>
       <exchange data="Forces" mesh="beam" from="Fluent" to="CSMdummy"/>
       <exchange data="Displacements" mesh="beam" from="CSMdummy" to="Fluent" initialize="0"/>
